% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reference_values.R
\docType{data}
\name{acceptable_edas}
\alias{acceptable_edas}
\title{English devolved areas lookup}
\format{
A data.frame with columns:
\describe{
\item{english_devolved_area_code}{GSS English devolved area code}
\item{english_devolved_area_name}{English devolved area name matching the
code}
}
}
\source{
dfe-published-data-qa screener app repo,
data/english_devolved_areas.csv
}
\usage{
acceptable_edas
}
\description{
A data frame of valid English devolved area code and name combinations as
used in EES data files.
}
\keyword{datasets}
