% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reference_values.R
\docType{data}
\name{acceptable_ethnicity_values}
\alias{acceptable_ethnicity_values}
\title{Acceptable ethnicity values}
\format{
A data.frame with columns:
\describe{
\item{ethnicity_major}{GSS ethnicity major group}
\item{ethnicity_minor}{GSS ethnicity minor group}
}
}
\source{
dfe-published-data-qa screener app repo, data/ethnicity.csv
}
\usage{
acceptable_ethnicity_values
}
\description{
A data frame of valid GSS ethnicity major and minor value pairs as used
in EES data files.
}
\keyword{datasets}
