% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reference_values.R
\docType{data}
\name{acceptable_extra_geog_options}
\alias{acceptable_extra_geog_options}
\title{Extra geography options}
\format{
A data.frame with columns:
\describe{
\item{code}{GSS geography code}
\item{name}{Geography name matching the code}
}
}
\source{
dfe-published-data-qa screener app repo,
data/universal_geog_options.csv
}
\usage{
acceptable_extra_geog_options
}
\description{
A data frame of universal geography code and name options that are
acceptable for any geography level (e.g. "Not applicable", "Unknown").
}
\keyword{datasets}
