% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reference_values.R
\docType{data}
\name{acceptable_indicator_units}
\alias{acceptable_indicator_units}
\title{Acceptable values for indicator units}
\format{
A vector of acceptable indicator unit values
}
\source{
maintained by explore.statistics@education.gov.uk
}
\usage{
acceptable_indicator_units
}
\description{
Acceptable values for indicator units
}
\keyword{datasets}
