% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reference_values.R
\docType{data}
\name{acceptable_lads}
\alias{acceptable_lads}
\title{Local authority districts lookup}
\format{
A data.frame with columns:
\describe{
\item{lad_code}{GSS LAD code}
\item{lad_name}{Local authority district name matching the code}
}
}
\source{
dfe-published-data-qa screener app repo, data/lads.csv
}
\usage{
acceptable_lads
}
\description{
A data frame of valid local authority district code and name combinations as
used in EES data files.
}
\keyword{datasets}
