% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reference_values.R
\docType{data}
\name{acceptable_leps}
\alias{acceptable_leps}
\title{Local enterprise partnerships lookup}
\format{
A data.frame with columns:
\describe{
\item{local_enterprise_partnership_code}{GSS LEP code}
\item{local_enterprise_partnership_name}{Local enterprise partnership name
matching the code}
}
}
\source{
dfe-published-data-qa screener app repo, data/leps.csv
}
\usage{
acceptable_leps
}
\description{
A data frame of valid local enterprise partnership code and name combinations
as used in EES data files.
}
\keyword{datasets}
