% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reference_values.R
\docType{data}
\name{acceptable_lsips}
\alias{acceptable_lsips}
\title{Local skills improvement plan areas lookup}
\format{
A data.frame with columns:
\describe{
\item{lsip_code}{GSS LSIP code}
\item{lsip_name}{Local skills improvement plan area name matching the code}
}
}
\source{
dfe-published-data-qa screener app repo, data/lsips.csv
}
\usage{
acceptable_lsips
}
\description{
A data frame of valid local skills improvement plan area code and name
combinations as used in EES data files. Includes both current and
historically renamed entries.
}
\keyword{datasets}
