% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reference_values.R
\docType{data}
\name{acceptable_regions}
\alias{acceptable_regions}
\title{Regions lookup}
\format{
A data.frame with columns:
\describe{
\item{region_code}{GSS region code}
\item{region_name}{Region name matching the code}
}
}
\source{
dfe-published-data-qa screener app repo, data/regions.csv
}
\usage{
acceptable_regions
}
\description{
A data frame of valid region codes and names as used in EES data files.
}
\keyword{datasets}
