% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/req_acceptable_values.R
\docType{data}
\name{acceptable_time_ids}
\alias{acceptable_time_ids}
\title{Acceptable time identifiers}
\format{
A vector of acceptable time identifier values
}
\source{
maintained by explore.statistics@education.gov.uk
}
\usage{
acceptable_time_ids
}
\description{
Acceptable time identifiers
}
\keyword{datasets}
