% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reference_values.R
\docType{data}
\name{acceptable_wards}
\alias{acceptable_wards}
\title{Wards lookup}
\format{
A data.frame with columns:
\describe{
\item{ward_code}{GSS ward code}
\item{ward_name}{Ward name matching the code}
}
}
\source{
dfe-published-data-qa screener app repo,
data/ward_lad_la_pcon_hierarchy.csv
}
\usage{
acceptable_wards
}
\description{
A data frame of valid ward codes and names as used in EES data files.
}
\keyword{datasets}
