% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reference_values.R
\docType{data}
\name{api_char_limits}
\alias{api_char_limits}
\title{API character limits}
\format{
A data.frame with columns:
\describe{
\item{name}{The name of the type of value}
\item{id}{The id of the type of value}
\item{char_limit}{The character limit for that type of value}
}
}
\source{
maintained by explore.statistics@education.gov.uk
}
\usage{
api_char_limits
}
\description{
API character limits
}
\keyword{datasets}
