% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_api_char_limit.R
\name{check_api_char_col_name}
\alias{check_api_char_col_name}
\title{Check if column names exceed a character limit}
\usage{
check_api_char_col_name(data, verbose = FALSE, stop_on_error = FALSE)
}
\arguments{
\item{data}{A data frame of the data file}

\item{verbose}{logical, if TRUE prints feedback messages to console for
every test, if FALSE run silently}

\item{stop_on_error}{logical, if TRUE will stop with an error if the result
is "FAIL", and will throw genuine warning if result is "WARNING"}
}
\value{
a single row data frame
}
\description{
Uses the \code{api_char_limit} function to check if any column names in the
provided data frame exceed the character limit specified in the
\code{api_char_limits} data frame.
}
\examples{
check_api_char_col_name(example_data)
}
\seealso{
Other check_api: 
\code{\link{check_api_char_col_label}()},
\code{\link{check_api_char_filter_items}()},
\code{\link{check_api_char_loc_code}()},
\code{\link{check_api_dict_col_names}()}
}
\concept{check_api}
