% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_api_char_limit.R
\name{check_api_char_filter_items}
\alias{check_api_char_filter_items}
\title{Check if filter items or location names exceed character limit}
\usage{
check_api_char_filter_items(data, meta, verbose = FALSE, stop_on_error = FALSE)
}
\arguments{
\item{data}{A data frame of the data file}

\item{meta}{A data frame of the metadata file}

\item{verbose}{logical, if TRUE prints feedback messages to console for
every test, if FALSE run silently}

\item{stop_on_error}{logical, if TRUE will stop with an error if the result
is "FAIL", and will throw genuine warning if result is "WARNING"}
}
\value{
a single row data frame
}
\description{
Checks if any values in filters, filter groups, or location names exceed the
character limit specified in the \code{api_char_limits} data frame, using the
\code{api_char_limit} function.
}
\examples{
check_api_char_filter_items(example_data, example_meta)
}
\seealso{
Other check_api: 
\code{\link{check_api_char_col_label}()},
\code{\link{check_api_char_col_name}()},
\code{\link{check_api_char_loc_code}()}
}
\concept{check_api}
