% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_api_char_limit.R
\name{check_api_char_limit}
\alias{check_api_char_limit}
\title{Check if values exceed a character limit}
\usage{
check_api_char_limit(values, type, verbose = FALSE, stop_on_error = FALSE)
}
\arguments{
\item{values}{A character vector of strings to check}

\item{type}{A character string specifying the type of values being checked,
one of "location-code", "column-name", "column-label" or "column-item".}

\item{verbose}{Logical, if TRUE prints feedback messages to console for
every test, if FALSE run silently}

\item{stop_on_error}{Logical, if TRUE will stop with an error if the result
is "FAIL", and will throw genuine warning if result is "WARNING"}
}
\value{
a single row data frame
}
\description{
Can be used to check individual strings, or vectors of strings.
}
\details{
Uses the exported \code{api_char_limits} data.frame as a reference for the
limits.
}
\examples{
check_api_char_limit(names(example_data), "column-name")
check_api_char_limit(names(example_data), "column-name", verbose = TRUE)
}
\seealso{
Other check_api: 
\code{\link{check_api_dict_col_names}()}
}
\concept{check_api}
