% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_api_dict_col_names.R
\name{check_api_dict_col_names}
\alias{check_api_dict_col_names}
\title{Check if col_names are present in the data dictionary}
\usage{
check_api_dict_col_names(meta, verbose = FALSE, stop_on_error = FALSE)
}
\arguments{
\item{meta}{dataframe containing a data set's meta data}

\item{verbose}{Logical, if TRUE prints feedback messages to console for
every test, if FALSE run silently}

\item{stop_on_error}{Logical, if TRUE will stop with an error if the result
is "FAIL", and will throw genuine warning if result is "WARNING"}
}
\description{
Uses the exported \code{data_dictionary} data.frame as a reference for the
acceptable col_names.
}
\examples{
check_api_dict_col_names(example_meta)
}
\seealso{
Other check_api: 
\code{\link{check_api_char_limit}()}
}
\concept{check_api}
