% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_col_ind_smushed.R
\name{check_col_ind_smushed}
\alias{check_col_ind_smushed}
\title{Check that no indicator column names contain typical filter entries}
\usage{
check_col_ind_smushed(meta, verbose = FALSE, stop_on_error = FALSE)
}
\arguments{
\item{meta}{A data frame of the metadata file}

\item{verbose}{logical, if TRUE prints feedback messages to console for
every test, if FALSE run silently}

\item{stop_on_error}{logical, if TRUE will stop with an error if the result
is "FAIL", and will throw genuine warning if result is "WARNING"}
}
\value{
a single row data frame
}
\description{
Checks the metadata file for any indicators that appear to be 'smushed'.
Flags any indicator col_name that contains common filter entries (e.g. male,
female, asian, black, etc), which suggests the data may not conform to tidy
data principles and should instead be pivoted longer with a filter column to
contain the characteristic choices.
}
\examples{
check_col_ind_smushed(example_meta)
check_col_ind_smushed(example_meta, verbose = TRUE)
}
\seealso{
Other check_col: 
\code{\link{check_col_snake_case}()},
\code{\link{check_col_var_start}()}
}
\concept{check_col}
