% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_col_names_spaces.R
\name{check_col_names_spaces}
\alias{check_col_names_spaces}
\title{Check for spaces in variable names
This function checks for spaces in the variable names of a given data frame.}
\usage{
check_col_names_spaces(data, verbose = FALSE, stop_on_error = FALSE)
}
\arguments{
\item{data}{A data frame of the data file}

\item{verbose}{logical, if TRUE prints feedback messages to console for
every test, if FALSE run silently}

\item{stop_on_error}{logical, if TRUE will stop with an error if the result
is "FAIL", and will throw genuine warning if result is "WARNING"}
}
\value{
a single row data frame
}
\description{
Check for spaces in variable names
This function checks for spaces in the variable names of a given data frame.
}
\examples{
check_col_names_spaces(example_data)
}
\concept{check_col_names}
