% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_col_snake_case.R
\name{check_col_snake_case}
\alias{check_col_snake_case}
\title{Check that column names follow snake_case convention}
\usage{
check_col_snake_case(data, verbose = FALSE, stop_on_error = FALSE)
}
\arguments{
\item{data}{A data frame to check for snake_case column names}

\item{verbose}{logical, if TRUE prints feedback messages to console for every test, if FALSE run silently}

\item{stop_on_error}{logical, if TRUE will stop with an error if the result is not "PASS"}
}
\value{
a single row data frame
}
\description{
This function checks if all column names in the data file follow the snake_case convention (lowercase letters, numbers, and underscores only).
}
\examples{
check_col_snake_case(example_data)
df <- example_data; names(df)[1] <- "SampleID"; check_col_snake_case(df)
}
\concept{check_col}
