% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_col_var_characteristic.R
\name{check_col_var_characteristic}
\alias{check_col_var_characteristic}
\title{Check for characteristic or characteristic_group variable names}
\usage{
check_col_var_characteristic(meta, verbose = FALSE, stop_on_error = FALSE)
}
\arguments{
\item{meta}{A data frame of the metadata file}

\item{verbose}{logical, if TRUE prints feedback messages to console for
every test, if FALSE run silently}

\item{stop_on_error}{logical, if TRUE will stop with an error if the result
is "FAIL", and will throw genuine warning if result is "WARNING"}
}
\value{
a single row data frame
}
\description{
Check whether the fields 'characteristic' or 'characteristic_group' have
been included in the metadata. These are not recommended for use with the
EES Table Tool.
}
\examples{
check_col_var_characteristic(example_meta)
check_col_var_characteristic(example_meta, verbose = TRUE)
}
\seealso{
Other check_col: 
\code{\link{check_col_snake_case}()},
\code{\link{check_col_var_start}()}
}
\concept{check_col}
