% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_col_var_start.R
\name{check_col_var_start}
\alias{check_col_var_start}
\title{Check that variable names start with a lowercase letter}
\usage{
check_col_var_start(data, verbose = FALSE, stop_on_error = FALSE)
}
\arguments{
\item{data}{A data frame of the data file}

\item{verbose}{logical, if TRUE prints feedback messages to console for
every test, if FALSE run silently}

\item{stop_on_error}{logical, if TRUE will stop with an error if the result
is "FAIL", and will throw genuine warning if result is "WARNING"}
}
\value{
a single row data frame
}
\description{
Check that all variable names in the data file start with a lowercase letter.
Flags any variables that start with a non-lowercase character as a warning.
}
\examples{
check_col_var_start(example_data)
check_col_var_start(example_data, verbose = TRUE)
}
\seealso{
Other check_col: 
\code{\link{check_col_ind_smushed}()},
\code{\link{check_col_snake_case}()}
}
\concept{check_col}
