% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_data_dictionary.R
\name{check_data_dict_col_name}
\alias{check_data_dict_col_name}
\title{Check col_names against the data dictionary}
\usage{
check_data_dict_col_name(meta, verbose = FALSE, stop_on_error = FALSE)
}
\arguments{
\item{meta}{A data frame of the metadata file}

\item{verbose}{logical, if TRUE prints feedback messages to console for
every test, if FALSE run silently}

\item{stop_on_error}{logical, if TRUE will stop with an error if the result
is "FAIL", and will throw genuine warning if result is "WARNING"}
}
\value{
a single row data frame
}
\description{
Validates that the filter and indicator column names in the metadata are
present in the data dictionary. Columns not in the data dictionary should
not be used in API data sets.
}
\examples{
check_data_dict_col_name(example_meta)
check_data_dict_col_name(example_meta, verbose = TRUE)
}
\seealso{
Other check_data_dict: 
\code{\link{check_data_dict_fil_item}()}
}
\concept{check_data_dict}
