% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_data_dictionary.R
\name{check_data_dict_fil_item}
\alias{check_data_dict_fil_item}
\title{Check filter items against the data dictionary}
\usage{
check_data_dict_fil_item(data, meta, verbose = FALSE, stop_on_error = FALSE)
}
\arguments{
\item{data}{A data frame of the data file}

\item{meta}{A data frame of the metadata file}

\item{verbose}{logical, if TRUE prints feedback messages to console for
every test, if FALSE run silently}

\item{stop_on_error}{logical, if TRUE will stop with an error if the result
is "FAIL", and will throw genuine warning if result is "WARNING"}
}
\value{
a single row data frame
}
\description{
Validates that the filter item values in the data are present in the data
dictionary for columns that are listed in the data dictionary. Filter items
not in the data dictionary should not be used in API data sets.
}
\details{
Only filter columns from the metadata that are also present in the data
dictionary are checked.
}
\examples{
check_data_dict_fil_item(example_data, example_meta)
check_data_dict_fil_item(example_data, example_meta, verbose = TRUE)
}
\seealso{
Other check_data_dict: 
\code{\link{check_data_dict_col_name}()}
}
\concept{check_data_dict}
