% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_empty_cols.R
\name{check_empty_cols2}
\alias{check_empty_cols2}
\title{Check for empty columns2}
\usage{
check_empty_cols2(df, filename = NULL, verbose = TRUE, custom_name = NULL)
}
\arguments{
\item{df}{The data frame to check}

\item{filename}{A character string of the filename to check}

\item{verbose}{Logical, if TRUE then messages are printed to the console,
otherwise a data frame is returned so multiple results can be combined into
a single table. Default is TRUE}

\item{custom_name}{An optional character string to customize the test name
in the output}
}
\value{
the output from \code{test_output()}
}
\description{
This function checks for empty cols in a given data frame and returns a list
with the check result and a message.
}
\details{
Filename is optional and is used to populate the feedback message.
}
\examples{
df <- eesyscreener::example_data
check_empty_cols2(df)
check_empty_cols2(df, verbose = FALSE)
}
\keyword{internal}
