% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_filename_spaces.R
\name{check_filename_spaces}
\alias{check_filename_spaces}
\title{Check for spaces in filename}
\usage{
check_filename_spaces(filename)
}
\arguments{
\item{filename}{A character string of the filename to check}
}
\value{
A list containing:
\describe{
\item{check}{Character string with the stage that the check belongs to.}
\item{result}{Character string, either "PASS", "FAIL" or "ADVISORY".}
\item{message}{Character string with feedback about the check.}
}
}
\description{
This function checks if the provided filename contains any spaces.
}
\examples{
check_filename_spaces("datafile.csv")
check_filename_spaces("data file.meta.csv")
}
