% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_filenames_match.R
\name{check_filenames_match}
\alias{check_filenames_match}
\title{Check filenames line up between data and metadata files}
\usage{
check_filenames_match(datafilename, metafilename, output = "console")
}
\arguments{
\item{datafilename}{A character string of the data filename to check}

\item{metafilename}{A character string of the metadata filename to check}

\item{output}{"table", "error-only", or "console". Default is "console"}
}
\value{
a console message with feedback from the check or a single row data
frame if verbose = FALSE
}
\description{
Making sure the data and metadata file follow the pattern of:
\itemize{
\item datafile.csv (data file)
\item datafile.meta.csv (metadata file)
}
}
\examples{
check_filenames_match("datafile.csv", "datafile.meta.csv")
check_filenames_match("datafile.csv", "metafile.csv", output = "table")
}
\seealso{
Other filename: 
\code{\link{check_filename_spaces}()},
\code{\link{check_filename_special}()}
}
\concept{filename}
