% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_filter_blanks.R
\name{check_filter_blanks}
\alias{check_filter_blanks}
\title{Check for blank values in filter and filter group columns}
\usage{
check_filter_blanks(data, meta, verbose = FALSE, stop_on_error = FALSE)
}
\arguments{
\item{data}{A data frame of the data file}

\item{meta}{A data frame of the metadata file}

\item{verbose}{logical, if TRUE prints feedback messages to console for
every test, if FALSE run silently}

\item{stop_on_error}{logical, if TRUE will stop with an error if the result
is "FAIL", and will throw genuine warning if result is "WARNING"}
}
\value{
a single row data frame
}
\description{
Checks that no filter or filter group columns contain blank (empty string)
values. Every cell in a filter or filter group column must have a value. If
a row represents no specific breakdown, such as 'all genders', the value
should be 'Total' rather than left blank.
}
\examples{
check_filter_blanks(example_data, example_meta)
check_filter_blanks(example_data, example_meta, verbose = TRUE)
}
\seealso{
Other check_filter: 
\code{\link{check_filter_defaults}()},
\code{\link{check_filter_group_level}()},
\code{\link{check_filter_item_limit}()},
\code{\link{check_filter_ob_total}()},
\code{\link{check_filter_whitespace}()}
}
\concept{check_filter}
