% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_filter_defaults.R
\name{check_filter_defaults}
\alias{check_filter_defaults}
\title{Check default filter values are present in data}
\usage{
check_filter_defaults(data, meta, verbose = FALSE, stop_on_error = FALSE)
}
\arguments{
\item{data}{A character string of the data filename to check}

\item{meta}{A character string of the metadata filename to check}

\item{verbose}{logical, if TRUE prints feedback messages to console for
every test, if FALSE run silently}

\item{stop_on_error}{logical, if TRUE will stop with an error if the result
is "FAIL", and will throw genuine warning if result is "WARNING"}
}
\value{
a single row data frame
}
\description{
Ensures that all filter columns and filter groups in the metadata have a
default filter value (usually "Total"), and that this value is present in
the corresponding columns of the data file.
}
\examples{
check_filter_defaults(example_data, example_meta)
}
\concept{check_filter}
