% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_filter_group_level.R
\name{check_filter_group_level}
\alias{check_filter_group_level}
\title{Check filter groups have an equal or lower number of levels}
\usage{
check_filter_group_level(data, meta, verbose = TRUE, stop_on_error = FALSE)
}
\arguments{
\item{data}{A data frame of the data file}

\item{meta}{A data frame of the metadata file}

\item{verbose}{logical, if TRUE prints feedback messages to console for
every test, if FALSE run silently}

\item{stop_on_error}{logical, if TRUE will stop with an error if the result
is "FAIL", and will throw genuine warning if result is "WARNING"}
}
\value{
a single row data frame
}
\description{
Ensure all filter groups have an equal or lower number of levels than their corresponding filter.
}
\examples{
check_filter_group_level(example_data, example_meta)
}
\concept{check_filter}
