% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_filter_item_limit.R
\name{check_filter_item_limit}
\alias{check_filter_item_limit}
\title{Check number of filter items are below limit}
\usage{
check_filter_item_limit(
  data,
  meta,
  filter_item_limit = 25000,
  verbose = FALSE,
  stop_on_error = FALSE
)
}
\arguments{
\item{data}{A character string of the data filename to check}

\item{meta}{A character string of the metadata filename to check}

\item{filter_item_limit}{The maximum number of unqiue items allowed in a single filter}

\item{verbose}{logical, if TRUE prints feedback messages to console for
every test, if FALSE run silently}

\item{stop_on_error}{logical, if TRUE will stop with an error if the result
is "FAIL", and will throw genuine warning if result is "WARNING"}
}
\value{
a single row data frame
}
\description{
Check number of filter items are below limit
}
\examples{
check_filter_item_limit(example_data, example_meta)
}
\seealso{
Other check_filter: 
\code{\link{check_filter_defaults}()},
\code{\link{check_filter_whitespace}()}
}
\concept{check_filter}
