% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_filter_whitespace.R
\name{check_filter_whitespace}
\alias{check_filter_whitespace}
\title{Check no filter values have leading or trailing whitespaces}
\usage{
check_filter_whitespace(data, meta, verbose = FALSE, stop_on_error = FALSE)
}
\arguments{
\item{data}{A data frame of the data file}

\item{meta}{A data frame of the metadata file}

\item{verbose}{logical, if TRUE prints feedback messages to console for
every test, if FALSE run silently}

\item{stop_on_error}{logical, if TRUE will stop with an error if the result
is "FAIL", and will throw genuine warning if result is "WARNING"}
}
\value{
a single row data frame
}
\description{
This function checks if the provided filter values contain any leading or trailing whitespaces.
}
\examples{
check_filter_whitespace(example_data, example_meta)
check_filter_whitespace(example_data, example_meta, verbose = TRUE)
}
\concept{check_filter}
