% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_general_null.R
\name{check_general_null}
\alias{check_general_null}
\title{Check for null and legacy no-data symbols}
\usage{
check_general_null(data, meta, verbose = FALSE, stop_on_error = FALSE)
}
\arguments{
\item{data}{A data frame of the data file}

\item{meta}{A data frame of the metadata file}

\item{verbose}{logical, if TRUE prints feedback messages to console for
every test, if FALSE run silently}

\item{stop_on_error}{logical, if TRUE will stop with an error if the result
is "FAIL", and will throw genuine warning if result is "WARNING"}
}
\value{
a single row data frame
}
\description{
Checks for null-like values (such as "NULL", "NA", and similar) in the
data and metadata files, and for legacy no-data symbols (such as "N/A",
".", "..") in the data file.
}
\details{
Null-like symbols cause a failure; finding legacy no-data symbols produces
a warning. The GSS provides guidance on the accepted symbols to use for
representing missing or suppressed data.
}
\examples{
check_general_null(example_data, example_meta)
}
\concept{check_general}
