% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_geog_country_combos.R
\name{check_geog_country_combos}
\alias{check_geog_country_combos}
\title{Check country code and name combinations}
\usage{
check_geog_country_combos(data, verbose = FALSE, stop_on_error = FALSE)
}
\arguments{
\item{data}{A data frame of the data file}

\item{verbose}{logical, if TRUE prints feedback messages to console for
every test, if FALSE run silently}

\item{stop_on_error}{logical, if TRUE will stop with an error if the result
is "FAIL", and will throw genuine warning if result is "WARNING"}
}
\value{
a single row data frame
}
\description{
Checks that all country_code and country_name combinations in the data file
are valid. Rows where country_code is "x" (the GSS not-available code) are
excluded from the check.
}
\examples{
check_geog_country_combos(example_data)
check_geog_country_combos(example_data, verbose = TRUE)
}
\concept{check_geog}
