% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_geog_ignored_rows.R
\name{check_geog_ignored_rows}
\alias{check_geog_ignored_rows}
\title{Check for rows ignored by the EES table tool}
\usage{
check_geog_ignored_rows(data, verbose = FALSE, stop_on_error = FALSE)
}
\arguments{
\item{data}{A data frame of the data file}

\item{verbose}{logical, if TRUE prints feedback messages to console for
every test, if FALSE run silently}

\item{stop_on_error}{logical, if TRUE will stop with an error if the result
is "FAIL", and will throw genuine warning if result is "WARNING"}
}
\value{
a single row data frame
}
\description{
Identifies rows at geographic levels that are ignored by the EES table tool:
School, Provider, Institution, and Planning area. Highlights in the message
if any such rows are present alongside other levels, and fails if:
\itemize{
\item School and Provider data have been mixed together
OR
\item The file only contains Planning area or Institution data
}
}
\examples{
check_geog_ignored_rows(example_data)
}
\seealso{
Other check_geog: 
\code{\link{check_geog_country_combos}()},
\code{\link{check_geog_eda_combos}()},
\code{\link{check_geog_la_combos}()},
\code{\link{check_geog_lad_combos}()},
\code{\link{check_geog_lep_combos}()},
\code{\link{check_geog_lsip_combos}()},
\code{\link{check_geog_other_dupes}()},
\code{\link{check_geog_pcon_combos}()},
\code{\link{check_geog_region_combos}()},
\code{\link{check_geog_region_for_la}()},
\code{\link{check_geog_region_for_lad}()},
\code{\link{check_geog_ward_combos}()}
}
\concept{check_geog}
