% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_geog_na.R
\name{check_geog_na}
\alias{check_geog_na}
\title{Check NA geography codes have the correct name}
\usage{
check_geog_na(data, verbose = FALSE, stop_on_error = FALSE)
}
\arguments{
\item{data}{A data frame of the data file}

\item{verbose}{logical, if TRUE prints feedback messages to console for
every test, if FALSE run silently}

\item{stop_on_error}{logical, if TRUE will stop with an error if the result
is "FAIL", and will throw genuine warning if result is "WARNING"}
}
\value{
a single row data frame
}
\description{
For geographic levels that have both a code and a name column, checks that
any location with the GSS not-available code \code{"x"} also has the
corresponding name \code{"Not available"}. Applies to all geographic levels
except those with their own combinations checks, RSC region, Institution,
and Planning area.
}
\examples{
check_geog_na(example_data)
check_geog_na(example_data, verbose = TRUE)
}
\seealso{
Other check_geog: 
\code{\link{check_geog_country_combos}()},
\code{\link{check_geog_eda_combos}()},
\code{\link{check_geog_ignored_rows}()},
\code{\link{check_geog_la_combos}()},
\code{\link{check_geog_lad_combos}()},
\code{\link{check_geog_lep_combos}()},
\code{\link{check_geog_lsip_combos}()},
\code{\link{check_geog_na_code}()},
\code{\link{check_geog_other_dupes}()},
\code{\link{check_geog_pcon_combos}()},
\code{\link{check_geog_region_col_present}()},
\code{\link{check_geog_region_combos}()},
\code{\link{check_geog_region_for_la}()},
\code{\link{check_geog_region_for_lad}()},
\code{\link{check_geog_ward_combos}()}
}
\concept{check_geog}
