% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_geog_region_for.R
\name{check_geog_region_for_la}
\alias{check_geog_region_for_la}
\title{Check that region columns are complete for Local authority rows}
\usage{
check_geog_region_for_la(data, verbose = FALSE, stop_on_error = FALSE)
}
\arguments{
\item{data}{A data frame of the data file}

\item{verbose}{logical, if TRUE prints feedback messages to console for
every test, if FALSE run silently}

\item{stop_on_error}{logical, if TRUE will stop with an error if the result
is "FAIL", and will throw genuine warning if result is "WARNING"}
}
\value{
a single row data frame
}
\description{
Checks that \code{region_code} and \code{region_name} columns are present in the data
file and contain no missing values for rows at the Local authority geographic
level. Regional information should ideally be provided for all Local
authority level data.
}
\details{
If no Local authority level data is present in the file, the check passes
immediately.
}
\examples{
check_geog_region_for_la(example_data)
}
\seealso{
Other check_geog: 
\code{\link{check_geog_country_combos}()},
\code{\link{check_geog_eda_combos}()},
\code{\link{check_geog_ignored_rows}()},
\code{\link{check_geog_la_combos}()},
\code{\link{check_geog_lad_combos}()},
\code{\link{check_geog_lep_combos}()},
\code{\link{check_geog_lsip_combos}()},
\code{\link{check_geog_na}()},
\code{\link{check_geog_na_code}()},
\code{\link{check_geog_other_dupes}()},
\code{\link{check_geog_pcon_combos}()},
\code{\link{check_geog_region_col_present}()},
\code{\link{check_geog_region_combos}()},
\code{\link{check_geog_region_for_lad}()},
\code{\link{check_geog_ward_combos}()}
}
\concept{check_geog}
