% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_geog_region_for.R
\name{check_geog_region_for_lad}
\alias{check_geog_region_for_lad}
\title{Check that region columns are complete for Local authority district rows}
\usage{
check_geog_region_for_lad(data, verbose = FALSE, stop_on_error = FALSE)
}
\arguments{
\item{data}{A data frame of the data file}

\item{verbose}{logical, if TRUE prints feedback messages to console for
every test, if FALSE run silently}

\item{stop_on_error}{logical, if TRUE will stop with an error if the result
is "FAIL", and will throw genuine warning if result is "WARNING"}
}
\value{
a single row data frame
}
\description{
Checks that \code{region_code} and \code{region_name} columns are present in the data
file and contain no missing values for rows at the Local authority district
geographic level. Regional information should ideally be provided for all
Local authority district level data.
}
\details{
If no Local authority district level data is present in the file, the check
passes immediately.
}
\examples{
check_geog_region_for_lad(example_data)
}
\seealso{
Other check_geog: 
\code{\link{check_geog_country_combos}()},
\code{\link{check_geog_ignored_rows}()},
\code{\link{check_geog_region_for_la}()}
}
\concept{check_geog}
