% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_geography_columns_present_and_expected.R
\name{check_geography_level_present}
\alias{check_geography_level_present}
\title{Check we have the right columns for the geographic level}
\usage{
check_geography_level_present(data, verbose = FALSE, stop_on_error = FALSE)
}
\arguments{
\item{data}{A data frame of the data file}

\item{verbose}{logical, if TRUE prints feedback messages to console for
every test, if FALSE run silently}

\item{stop_on_error}{logical, if TRUE will stop with an error if the result
is "FAIL", and will throw genuine warning if result is "WARNING"}
}
\description{
Check we have the right columns for the geographic level
}
\examples{
check_geography_level_present(example_data)
check_geography_level_present(example_data, verbose = TRUE)
check_geography_level_present(example_comma_data, verbose = TRUE)
}
\concept{check_geography}
