% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_harmonised.R
\name{check_harmonised_variables}
\alias{check_harmonised_variables}
\title{Check col names against harmonised data standards}
\usage{
check_harmonised_variables(meta, verbose = FALSE, stop_on_error = FALSE)
}
\arguments{
\item{meta}{A data frame of the metadata file}

\item{verbose}{logical, if TRUE prints feedback messages to console for
every test, if FALSE run silently}

\item{stop_on_error}{logical, if TRUE will stop with an error if the result
is "FAIL", and will throw genuine warning if result is "WARNING"}
}
\value{
a single row data frame
}
\description{
Validates that the column names and filter grouping column names in the
metadata are not using non-standard names for variables covered by the DfE
harmonised data standards. Any column name that matches a harmonised variable
search pattern but is not one of the accepted standard names will fail.
}
\examples{
check_harmonised_variables(example_meta)
check_harmonised_variables(example_meta, verbose = TRUE)
}
\seealso{
Other check_harmonised: 
\code{\link{check_harmonised_eth_char_grp}()},
\code{\link{check_harmonised_eth_char_vals}()},
\code{\link{check_harmonised_eth_vals}()}
}
\concept{check_harmonised}
