% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_ind_invalid_entry.R
\name{check_ind_invalid_entry}
\alias{check_ind_invalid_entry}
\title{Check for invalid values in indicators}
\usage{
check_ind_invalid_entry(data, meta, verbose = FALSE, stop_on_error = FALSE)
}
\arguments{
\item{data}{A data frame of the data file}

\item{meta}{A data frame of the metadata file}

\item{verbose}{logical, if TRUE prints feedback messages to console for
every test, if FALSE run silently}

\item{stop_on_error}{logical, if TRUE will stop with an error if the result
is "FAIL", and will throw genuine warning if result is "WARNING"}
}
\value{
a single row data frame
}
\description{
Check all values in indicator columns for blanks or obsolete no data symbols.
}
\examples{
check_ind_invalid_entry(example_data, example_meta)
check_ind_invalid_entry(example_data, example_meta, verbose = TRUE)
}
\concept{check_ind}
