% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_ind_obsolete_symbols.R
\name{check_ind_obsolete_symbols}
\alias{check_ind_obsolete_symbols}
\title{Check that there are no obsolete symbols (i.e.  ~, : ) in any indicators.}
\usage{
check_ind_obsolete_symbols(data, meta, verbose = FALSE, stop_on_error = FALSE)
}
\arguments{
\item{data}{A data frame of the data file}

\item{meta}{A data frame of the metadata file}

\item{verbose}{logical, if TRUE prints feedback messages to console for
every test, if FALSE run silently}

\item{stop_on_error}{logical, if TRUE will stop with an error if the result
is "FAIL", and will throw genuine warning if result is "WARNING"}
}
\value{
a single row data frame
}
\description{
Throw an error if any obsolete symbols (i.e.  ~, : ), are present in the indicator values.
}
\examples{
check_ind_obsolete_symbols(example_data, example_meta)
check_ind_obsolete_symbols(example_data, example_meta, verbose = TRUE)
}
\seealso{
Other check_ind: 
\code{\link{check_ind_blanks}()}
}
\concept{check_ind}
