% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_meta_col_name_duplicate.R
\name{check_meta_col_name_duplicate}
\alias{check_meta_col_name_duplicate}
\title{Check there are no duplicated column names}
\usage{
check_meta_col_name_duplicate(meta, verbose = FALSE, stop_on_error = FALSE)
}
\arguments{
\item{meta}{A data frame of the metadata file}

\item{verbose}{logical, if TRUE prints feedback messages to console for
every test, if FALSE run silently}

\item{stop_on_error}{logical, if TRUE will stop with an error if the result
is "FAIL", and will throw genuine warning if result is "WARNING"}
}
\value{
a single row data frame
}
\description{
Make sure that all column names within the \code{col_name} column of the metadata file are unique.
}
\examples{
check_meta_col_name_duplicate(example_meta)
check_meta_col_name_duplicate(example_meta, verbose = TRUE)
}
\seealso{
Other check_meta: 
\code{\link{check_meta_duplicate_label}()},
\code{\link{check_meta_filter_group}()},
\code{\link{check_meta_filter_group_match}()},
\code{\link{check_meta_filter_hint}()},
\code{\link{check_meta_ind_dp_set}()},
\code{\link{check_meta_indicator_dp}()},
\code{\link{check_meta_label}()},
\code{\link{check_meta_whitespace_filters}()}
}
\concept{check_meta}
