% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_meta_fil_grp_dupe.R
\name{check_meta_fil_grp_dupe}
\alias{check_meta_fil_grp_dupe}
\title{Check all of the filter_group values are unique}
\usage{
check_meta_fil_grp_dupe(meta, verbose = FALSE, stop_on_error = FALSE)
}
\arguments{
\item{meta}{A data frame of the metadata file}

\item{verbose}{logical, if TRUE prints feedback messages to console for
every test, if FALSE run silently}

\item{stop_on_error}{logical, if TRUE will stop with an error if the result
is "FAIL", and will throw genuine warning if result is "WARNING"}
}
\value{
a single row data frame
}
\description{
Ensure all filter groups from the meta data are unique.
}
\examples{
check_meta_fil_grp_dupe(example_meta)
check_meta_fil_grp_dupe(example_meta, verbose = TRUE)
}
\seealso{
Other check_meta: 
\code{\link{check_meta_col_name_dupe}()},
\code{\link{check_meta_dupe_label}()},
\code{\link{check_meta_fil_grp}()},
\code{\link{check_meta_fil_grp_is_fil}()},
\code{\link{check_meta_fil_grp_match}()},
\code{\link{check_meta_fil_grp_stripped}()},
\code{\link{check_meta_filter_hint}()},
\code{\link{check_meta_geog_catch}()},
\code{\link{check_meta_ind_dp_set}()},
\code{\link{check_meta_ind_dp_values}()},
\code{\link{check_meta_ind_unit}()},
\code{\link{check_meta_ind_unit_validation}()},
\code{\link{check_meta_indicator_dp}()},
\code{\link{check_meta_indicator_grouping}()},
\code{\link{check_meta_label}()}
}
\concept{check_meta}
