% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_meta_filter_group.R
\name{check_meta_filter_group}
\alias{check_meta_filter_group}
\title{Check filter_group is not set for indicator rows}
\usage{
check_meta_filter_group(meta, verbose = FALSE, stop_on_error = FALSE)
}
\arguments{
\item{meta}{A data frame of the metadata file}

\item{verbose}{logical, if TRUE prints feedback messages to console for
every test, if FALSE run silently}

\item{stop_on_error}{logical, if TRUE will stop with an error if the result
is "FAIL", and will throw genuine warning if result is "WARNING"}
}
\value{
a single row data frame
}
\description{
Throw warning if there are any indicator rows that have filter_group as
non-blank, to encourage users to correct the metadata.
}
\examples{
check_meta_filter_group(example_meta)
check_meta_filter_group(example_meta, verbose = TRUE)
}
\seealso{
Other check_meta: 
\code{\link{check_meta_col_name_duplicate}()},
\code{\link{check_meta_duplicate_label}()},
\code{\link{check_meta_filter_group_is_filter}()},
\code{\link{check_meta_filter_group_match}()},
\code{\link{check_meta_filter_group_stripped}()},
\code{\link{check_meta_filter_hint}()},
\code{\link{check_meta_ind_dp_set}()},
\code{\link{check_meta_indicator_dp}()},
\code{\link{check_meta_indicator_grouping}()},
\code{\link{check_meta_label}()}
}
\concept{check_meta}
