% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_meta_filter_hint.R
\name{check_meta_filter_hint}
\alias{check_meta_filter_hint}
\title{Check filter_hint is not set for indicator rows}
\usage{
check_meta_filter_hint(meta, verbose = FALSE, stop_on_error = FALSE)
}
\arguments{
\item{meta}{A data frame of the metadata file}

\item{verbose}{logical, if TRUE prints feedback messages to console for
every test, if FALSE run silently}

\item{stop_on_error}{logical, if TRUE will stop with an error if the result
is "FAIL", and will throw genuine warning if result is "WARNING"}
}
\value{
a single row data frame
}
\description{
Throw warning if there are any indicator rows that have filter_hint as
non-blank, to encourage users to correct the metadata.
}
\examples{
check_meta_filter_hint(example_meta)
check_meta_filter_hint(example_meta, verbose = TRUE)
}
\seealso{
Other check_meta: 
\code{\link{check_meta_ind_dp_set}()}
}
\concept{check_meta}
