% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_meta_label.R
\name{check_meta_label}
\alias{check_meta_label}
\title{Check every row has a label}
\usage{
check_meta_label(meta, verbose = FALSE, stop_on_error = FALSE)
}
\arguments{
\item{meta}{A data frame of the metadata file}

\item{verbose}{logical, if TRUE prints feedback messages to console for
every test, if FALSE run silently}

\item{stop_on_error}{logical, if TRUE will stop with an error if the result
is "FAIL", and will throw genuine warning if result is "WARNING"}
}
\value{
a single row data frame
}
\description{
Ensure there are no blank cells in the label column for every row in the metafile
}
\examples{
check_meta_label(example_meta)
check_meta_label(example_meta, verbose = TRUE)
}
\seealso{
Other check_meta: 
\code{\link{check_meta_ind_dp_set}()}
}
\concept{check_meta}
