% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_meta_whitespace_filters.R
\name{check_meta_whitespace_filters}
\alias{check_meta_whitespace_filters}
\title{Check no filter labels have leading or trailing whitespaces}
\usage{
check_meta_whitespace_filters(
  data,
  meta,
  verbose = FALSE,
  stop_on_error = FALSE
)
}
\arguments{
\item{data}{A data frame of the data file}

\item{meta}{A data frame of the metadata file}

\item{verbose}{logical, if TRUE prints feedback messages to console for
every test, if FALSE run silently}

\item{stop_on_error}{logical, if TRUE will stop with an error if the result
is "FAIL", and will throw genuine warning if result is "WARNING"}
}
\value{
a single row data frame
}
\description{
This function checks if the provided filter labels contain any leading or trailing whitespaces.
}
\examples{
check_meta_whitespace_filters(example_data, example_meta)
check_meta_whitespace_filters(example_data, example_meta, verbose = TRUE)
}
\seealso{
Other check_meta: 
\code{\link{check_meta_col_name_duplicate}()},
\code{\link{check_meta_duplicate_label}()},
\code{\link{check_meta_filter_group}()},
\code{\link{check_meta_filter_group_match}()},
\code{\link{check_meta_filter_hint}()},
\code{\link{check_meta_ind_dp_set}()},
\code{\link{check_meta_indicator_dp}()},
\code{\link{check_meta_label}()}
}
\concept{check_meta}
