% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_time_period.R
\name{check_time_period}
\alias{check_time_period}
\title{Check that time periods match the time identifier}
\usage{
check_time_period(data, verbose = FALSE, stop_on_error = FALSE)
}
\arguments{
\item{data}{A data frame of the data file}

\item{verbose}{logical, if TRUE prints feedback messages to console for
every test, if FALSE run silently}

\item{stop_on_error}{logical, if TRUE will stop with an error if the result
is "FAIL", and will throw genuine warning if result is "WARNING"}
}
\value{
a single row data frame
}
\description{
Every time identifier is only compatible with either 4 or 6 digit time
periods.
}
\examples{
check_time_period(example_data)
}
\seealso{
Other check_time: 
\code{\link{check_time_period_six}()}
}
\concept{check_time}
