% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_time_period_six.R
\name{check_time_period_six}
\alias{check_time_period_six}
\title{Check that 6 digit time periods give consecutive years}
\usage{
check_time_period_six(data, verbose = FALSE, stop_on_error = FALSE)
}
\arguments{
\item{data}{A data frame of the data file}

\item{verbose}{logical, if TRUE prints feedback messages to console for
every test, if FALSE run silently}

\item{stop_on_error}{logical, if TRUE will stop with an error if the result
is "FAIL", and will throw genuine warning if result is "WARNING"}
}
\value{
a single row data frame
}
\description{
Every 6 digit time indentifier should be in the format YYYYZZ where YYYY is
a 4 digit year and ZZ is the final two digits of the following year.
}
\details{
For example, 202122 is valid but 202120 is not.
}
\examples{
check_time_period_six(example_data)
}
\seealso{
Other check_time: 
\code{\link{check_time_period}()}
}
\concept{check_time}
