% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reference_values.R
\docType{data}
\name{countries}
\alias{countries}
\title{Countries lookup}
\format{
A data.frame with columns:
\describe{
\item{country_code}{GSS country code}
\item{country_name}{Country name matching the code}
}
}
\source{
dfe-published-data-qa screener app repo, data/country.csv
}
\usage{
countries
}
\description{
A data frame of valid country codes and names as used in EES data files.
}
\keyword{datasets}
