% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reference_values.R
\docType{data}
\name{data_dictionary}
\alias{data_dictionary}
\title{Data dictionary}
\format{
A data frame with columns:
\describe{
\item{col_name}{The name of the column as it should appear in the data file}
\item{col_type}{The type of column (e.g. Indicator, Filter)}
\item{filter_item}{Acceptable values for filter columns}
\item{filter_item_parent}{The parent value for the filter item if used in a hierarchy}
\item{col_name_parent}{The parent column name for the column if used in a hierarchy}
\item{cbds_code}{The code for the value in the Common Basic Data Set}
\item{item_error_flag}{How the item is used in screening checks}
\item{contexts}{A rough grouping to aid navigation of the data dictionary}
}
}
\source{
maintained by explore.statistics@education.gov.uk
}
\usage{
data_dictionary
}
\description{
A data frame containing the data dictionary for data sets, including
column names and acceptable values within those columns.
}
\keyword{datasets}
