% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_datasets.R
\docType{data}
\name{ees_robot_test_data}
\alias{ees_robot_test_data}
\title{Example screening output pre-generated from example_data and example_meta}
\format{
A list of 14 robot test examples, with each element of the list containing
\describe{
\item{robot name}{Name of the robot test data}
\item{data}{The robot test data}
\item{meta}{The robot test meta}
}
}
\source{
maintained by explore.statistics@education.gov.uk
}
\usage{
ees_robot_test_data
}
\description{
Example screening output pre-generated from example_data and example_meta
}
\keyword{datasets}
