% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eesyscreener-package.R
\docType{package}
\name{eesyscreener-package}
\alias{eesyscreener}
\alias{eesyscreener-package}
\title{eesyscreener: Explore education statistics data screening}
\description{
Quality assurance checks used to enforce standards of official statistics files ready for publication on the explore education statistics platform.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://dfe-analytical-services.github.io/eesyscreener/}
  \item \url{https://github.com/dfe-analytical-services/eesyscreener}
  \item Report bugs at \url{https://github.com/dfe-analytical-services/eesyscreener/issues}
}

}
\author{
\strong{Maintainer}: Cam Race \email{cameron.race@education.gov.uk}

Authors:
\itemize{
  \item Rich Bielby \email{richard.bielby@education.gov.uk} (\href{https://orcid.org/0000-0001-9070-9969}{ORCID})
}

Other contributors:
\itemize{
  \item Charlotte Foster \email{charlotte.foster@education.gov.uk} [contributor]
  \item Lara Garbett \email{lara.garbett@education.gov.uk} [contributor]
  \item Pauline Potts \email{pauline.potts@education.gov.uk} [contributor]
  \item Matthew Robinson \email{matthew3.robinson@education.gov.uk} [contributor]
  \item Ethan Roy \email{ethan.roy@education.gov.uk} [contributor]
  \item Lauren Snaathorst \email{lauren.snaathorst@education.gov.uk} [contributor]
  \item Bob Tarrant \email{robert.tarrant@education.gov.uk} [contributor]
  \item Chris Hall \email{christopher.hall@education.gov.uk} [contributor]
  \item Department for Education, England \email{explore.statistics@education.gov.uk} [copyright holder]
}

}
\keyword{internal}
