% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_datasets.R
\docType{data}
\name{example_api_long}
\alias{example_api_long}
\title{Example data that has a column name too long for the API}
\format{
A data frame with 3 rows and 8 columns:
\describe{
\item{time_period}{Time period of row}
\item{time_identifier}{Time identifier of row}
\item{geographic_level}{Geographic level of row}
\item{country_code}{Country code (9 digits)}
\item{country_name}{Country name}
\item{establishment_type_group}{Type of establishment}
\item{mahoooooooooooooooooooooooooooooooooooooooooooooooooooosive}{Long}
\item{sex}{Sex of pupil}
\item{enrolment_count}{Number of enrolments}
}
}
\source{
maintained by explore.statistics@education.gov.uk
}
\usage{
example_api_long
}
\description{
Example data that has a column name too long for the API
}
\keyword{datasets}
