% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_data.R
\docType{data}
\name{example_data}
\alias{example_data}
\title{Example data}
\format{
A data frame with 9 rows and 8 columns:
\describe{
\item{time_period}{Time period of row}
\item{time_identifier}{Time identifier of row}
\item{geographic_level}{Geographic level of row}
\item{country_code}{Country code (9 digits)}
\item{country_name}{Country name}
\item{sex}{Sex of pupil}
\item{establishment_type}{Type of establishment}
\item{enrolment_count}{Number of enrolments}
}
}
\source{
maintained by explore.statistics@education.gov.uk
}
\usage{
example_data
}
\description{
Simple example data file.
}
\keyword{datasets}
