% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/_example_datasets.R
\docType{data}
\name{example_filter_group}
\alias{example_filter_group}
\title{Example filter group data}
\format{
A data frame with 9 rows and 8 columns:
\describe{
\item{time_period}{Time period of row}
\item{time_identifier}{Time identifier of row}
\item{geographic_level}{Geographic level of row}
\item{country_code}{Country code (9 digits)}
\item{country_name}{Country name}
\item{establishment_type_group}{Type of establishment}
\item{education_phase}{Education phase}
\item{enrolment_count}{Number of enrolments}
}
}
\source{
maintained by explore.statistics@education.gov.uk
}
\usage{
example_filter_group
}
\description{
Simple example data file showing an example of filter groups in action
}
\keyword{datasets}
