% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_data.R
\docType{data}
\name{example_meta}
\alias{example_meta}
\title{Example metadata}
\format{
A data frame with 3 rows and 8 columns:
\describe{
\item{col_name}{Name of column in data file}
\item{col_type}{Type of colummn in data file}
\item{label}{Human readable label for column}
\item{indicator_grouping}{Grouping for indicator columns}
\item{indicator_unit}{Unit prefix or suffix}
\item{indicator_dp}{Number of decimal places}
\item{filter_hint}{Optional hint text to accompany filter}
\item{filter_grouping_column}{Grouping column showing a filter hierachy}
}
}
\source{
maintained by explore.statistics@education.gov.uk
}
\usage{
example_meta
}
\description{
This data set contains example metadata for the example data.
}
\keyword{datasets}
