% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_datasets.R
\docType{data}
\name{example_output}
\alias{example_output}
\title{Example screening output pre-generated from example_data and example_meta}
\format{
A data frame with 21 rows and 5 columns:
\describe{
\item{check}{Name of the check}
\item{result}{Check result: PASS / FAIL / WARNING}
\item{message}{Check info message}
\item{guidance_url}{URL to any guidance associated with the check}
\item{stage}{Stage group the listed check belongs to}
}
}
\source{
maintained by explore.statistics@education.gov.uk
}
\usage{
example_output
}
\description{
Example screening output pre-generated from example_data and example_meta
}
\keyword{datasets}
