% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stage_wrappers.R
\name{fileValidation}
\alias{fileValidation}
\title{Validate Data and Metadata Files}
\usage{
fileValidation(
  datafilename,
  metafilename,
  dataseparator,
  metaseparator,
  datafile,
  metafile
)
}
\arguments{
\item{datafilename}{A string representing the name of the data file.}

\item{metafilename}{A string representing the name of the metadata file.}

\item{dataseparator}{A string representing the separator used in the data file.}

\item{metaseparator}{A string representing the separator used in the metadata file.}

\item{datafile}{A data frame containing the data file content.}

\item{metafile}{A data frame containing the metadata file content.}
}
\value{
A tibble with the results of the validation checks.
}
\description{
This function performs a series of validation checks on data and metadata files.
}
\details{
The function performs the following validation checks:
\itemize{
  \item Checks for spaces in the data and metadata filenames.
  \item Checks for special characters in the data and metadata filenames.
  \item Ensures naming conventions are followed for the data and metadata filenames.
  \item Converts rows to columns for comparison between data and metadata files.
  \item Validates the file separators used in the data and metadata files.
  \item Checks for empty rows in the data and metadata files.
  \item Checks for empty columns in the data file.
  \item Ensures mandatory columns are present in the data and metadata files.
}
}
\examples{
\dontrun{
datafile <- read.csv("data.csv")
metafile <- read.csv("meta.csv")
result <- fileValidation("data.csv", "meta.csv", ",", ",", datafile, metafile)
print(result)
}

}
