% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reference_values.R
\docType{data}
\name{filter_item_limit}
\alias{filter_item_limit}
\title{Filter item limit}
\format{
A numeric giving the maximum number of allowed distinct filter entries
}
\source{
maintained by explore.statistics@education.gov.uk
}
\usage{
filter_item_limit
}
\description{
Filter item limit
}
\keyword{datasets}
