% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_duplicate_rows.R
\name{find_duplicate_rows}
\alias{find_duplicate_rows}
\title{Identify duplicate rows}
\usage{
find_duplicate_rows(data, meta)
}
\arguments{
\item{data}{A data frame of the data file}
}
\description{
Identify duplicate rows in a data set
}
\examples{
find_duplicate_rows(bind_rows(example_data, example_data), example_meta)
}
