% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reference_values.R
\docType{data}
\name{four_digit_identifiers}
\alias{four_digit_identifiers}
\title{Time identifiers that should have 4 digit numbers}
\format{
A vector of time identifier values that should have 4 digit numbers
}
\source{
maintained by explore.statistics@education.gov.uk
}
\usage{
four_digit_identifiers
}
\description{
A vector of time identifier values that should have 4 digit numbers.
This is a subset of the acceptable_time_ids list.
}
\keyword{datasets}
