% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_test_files.R
\name{generate_meta_file}
\alias{generate_meta_file}
\title{Generate a meta file for testing}
\usage{
generate_meta_file(num_filters, num_indicators)
}
\arguments{
\item{num_filters}{The number of filters to generate}

\item{num_indicators}{The number of indicators to generate}
}
\value{
A data frame following the EES metadata standards
}
\description{
This function is designed to be used within \code{generate_test_dfs} to
complement the data file generated by \code{generate_data_file}.
}
\keyword{internal}
