% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_test_files.R
\name{generate_test_dfs}
\alias{generate_test_dfs}
\title{EES-ily generate some beefy test files}
\usage{
generate_test_dfs(
  years = 3000,
  pcon_names = "Sheffield Central",
  pcon_codes = "E14000919",
  num_filters = 1,
  num_indicators = 1
)
}
\arguments{
\item{years}{A vector of 4 digit years}

\item{pcon_names}{A vector of PCon names}

\item{pcon_codes}{A vector of PCon codes}

\item{num_filters}{The number of filters to generate, currently
will give each filter 8 items}

\item{num_indicators}{The number of indicators to generate, currently
generates random numbers between 100 and 1000}
}
\value{
A list containing two data frames, one data and associated meta
}
\description{
Mostly for ease of performance benchmarking. A bit crude, only does
PCons for simplicity so far.
}
\examples{
generate_test_dfs(2010:2015, "Sheffield Central", "E14000919", 2, 3)
}
