% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/req_acceptable_values.R
\docType{data}
\name{geography_df}
\alias{geography_df}
\title{Acceptable geographic levels and their associated columns}
\format{
data.frame
}
\source{
maintained by explore.statistics@education.gov.uk
}
\usage{
geography_df
}
\description{
Acceptable geographic levels and their associated columns
}
\keyword{datasets}
