% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reference_values.R
\docType{data}
\name{gss_symbols}
\alias{gss_symbols}
\title{Acceptable values for GSS symbols}
\format{
A vector of acceptable GSS symbols
}
\source{
maintained by explore.statistics@education.gov.uk
}
\usage{
gss_symbols
}
\description{
Acceptable values for GSS symbols
}
\keyword{datasets}
