% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reference_values.R
\docType{data}
\name{legacy_gss_symbols}
\alias{legacy_gss_symbols}
\title{Legacy values for GSS symbols}
\format{
A vector of legacy GSS symbols
}
\source{
maintained by explore.statistics@education.gov.uk
}
\usage{
legacy_gss_symbols
}
\description{
Legacy values for GSS symbols
}
\keyword{datasets}
