% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{null_filename}
\alias{null_filename}
\title{Handle null filenames}
\usage{
null_filename(string = NULL)
}
\arguments{
\item{string}{filename in to be transformed}
}
\value{
string
}
\description{
Helper function to handle null filenames and return a string for the
message.
}
\keyword{internal}
