% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reference_values.R
\docType{data}
\name{optional_meta_cols}
\alias{optional_meta_cols}
\title{Optional metadata columns}
\format{
A vector of the acceptable, optional metadata column names
}
\source{
maintained by explore.statistics@education.gov.uk
}
\usage{
optional_meta_cols
}
\description{
Optional metadata columns
}
\keyword{datasets}
