% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reference_values.R
\docType{data}
\name{potential_ob_units_regex}
\alias{potential_ob_units_regex}
\title{Regex pattern for potential observational unit columns}
\format{
A length 1 character vector containing a regex pattern
}
\source{
maintained by explore.statistics@education.gov.uk
}
\usage{
potential_ob_units_regex
}
\description{
A regex pattern that matches common geography-related and other observational
unit column names (e.g. names containing prefixes like "la", "reg", "pcon"
or suffixes like "name", "code", "urn", "ukprn").
}
\keyword{datasets}
