% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/precheck_col_invalid_meta.R
\name{precheck_col_invalid_meta}
\alias{precheck_col_invalid_meta}
\title{Check there are no invalid columns in the metadata}
\usage{
precheck_col_invalid_meta(meta, output = "console")
}
\arguments{
\item{meta}{A data frame of the metadata file}

\item{output}{"table", "error-only", or "console". Default is "console"}
}
\value{
a console message with feedback from the check or a single row data
frame if verbose = FALSE
}
\description{
Make sure that all column names are expected metadata columns. Checks across
both \code{req_meta_cols} and \code{optional_meta_cols}.
}
\examples{
precheck_col_invalid_meta(example_meta)
precheck_col_invalid_meta(example_meta, output = "table")
}
\seealso{
Other precheck_col: 
\code{\link{precheck_col_req_meta}()}
}
\concept{precheck_col}
