% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/precheck_col_name_duplicate.R
\name{precheck_col_name_duplicate}
\alias{precheck_col_name_duplicate}
\title{Check there are no duplicated column names}
\usage{
precheck_col_name_duplicate(meta, verbose = FALSE, stop_on_error = FALSE)
}
\arguments{
\item{meta}{A data frame of the metadata file}

\item{verbose}{logical, if TRUE prints feedback messages to console for
every test, if FALSE run silently}

\item{stop_on_error}{logical, if TRUE will stop with an error if the result
is "FAIL", and will throw genuine warning if result is "WARNING"}
}
\value{
a single row data frame
}
\description{
Make sure that all column names are unique.
}
\examples{
precheck_col_name_duplicate(example_meta)
precheck_col_name_duplicate(example_meta, verbose = TRUE)
}
\seealso{
Other precheck_col: 
\code{\link{precheck_col_invalid_meta}()},
\code{\link{precheck_col_req_data}()},
\code{\link{precheck_col_req_meta}()},
\code{\link{precheck_col_to_rows}()}
}
\concept{precheck_col}
