% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/precheck_col_req_data.R
\name{precheck_col_req_data}
\alias{precheck_col_req_data}
\title{Check all required columns are present in data}
\usage{
precheck_col_req_data(data, output = "console")
}
\arguments{
\item{data}{A data frame of the data file}

\item{output}{"table", "error-only", or "console". Default is "console"}
}
\value{
a console message with feedback from the check or a single row data
frame if verbose = FALSE
}
\description{
Using the \code{req_data_cols} object.
}
\examples{
precheck_col_req_data(example_data)
precheck_col_req_data(example_data, output = "table")
}
\seealso{
Other precheck_col: 
\code{\link{precheck_col_invalid_meta}()},
\code{\link{precheck_col_req_meta}()},
\code{\link{precheck_col_to_rows}()}
}
\concept{precheck_col}
