% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/precheck_col_req_meta.R
\name{precheck_col_req_meta}
\alias{precheck_col_req_meta}
\title{Check all required columns are present in metadata}
\usage{
precheck_col_req_meta(meta, output = "console")
}
\value{
a console message with feedback from the check or a single row data
frame if verbose = FALSE
}
\description{
Using the \code{mandatory_meta_cols} object.
}
\examples{
precheck_col_req_meta(example_meta)
precheck_col_req_meta(example_meta, output = "table")
}
\concept{precheck_col}
