% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/precheck_cross_data_to_meta.R
\name{precheck_cross_data_to_meta}
\alias{precheck_cross_data_to_meta}
\title{Check all data variables exist in the metadata file}
\usage{
precheck_cross_data_to_meta(data, meta, verbose = FALSE, stop_on_error = FALSE)
}
\arguments{
\item{data}{A data frame of the data file}

\item{meta}{A data frame of the metadata file}

\item{verbose}{logical, if TRUE prints feedback messages to console for
every test, if FALSE run silently}

\item{stop_on_error}{logical, if TRUE will stop with an error if the result
is "FAIL", and will throw genuine warning if result is "WARNING"}
}
\value{
a single row data frame
}
\description{
Ensure that every variable in the data file is either an observational unit,
has only a single unique value, or is represented in the metadata file as
either a \code{col_name} or \code{filter_grouping_column} entry.
}
\examples{
precheck_cross_data_to_meta(example_data, example_meta)
precheck_cross_data_to_meta(example_data, example_meta, verbose = TRUE)
}
\seealso{
Other precheck_cross: 
\code{\link{precheck_cross_meta_to_data}()}
}
\concept{precheck_cross}
