% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/precheck_cross_meta_to_data.R
\name{precheck_cross_meta_to_data}
\alias{precheck_cross_meta_to_data}
\title{Check all metadata variables exist in the data file}
\usage{
precheck_cross_meta_to_data(data, meta, verbose = FALSE, stop_on_error = FALSE)
}
\arguments{
\item{data}{A data frame of the data file}

\item{meta}{A data frame of the metadata file}

\item{verbose}{logical, if TRUE prints feedback messages to console for
every test, if FALSE run silently}

\item{stop_on_error}{logical, if TRUE will stop with an error if the result
is "FAIL", and will throw genuine warning if result is "WARNING"}
}
\value{
a single row data frame
}
\description{
Ensure that every variable named in the metadata file can be found as a
column in the data file. This covers both \code{col_name} entries and any
\code{filter_grouping_column} values.
}
\examples{
precheck_cross_meta_to_data(example_data, example_meta)
precheck_cross_meta_to_data(example_data, example_meta, verbose = TRUE)
}
\concept{precheck_cross}
