% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/precheck_filter_not_singular.R
\name{precheck_filter_not_singular}
\alias{precheck_filter_not_singular}
\title{Check all filters have more than one level}
\usage{
precheck_filter_not_singular(
  data,
  meta,
  verbose = FALSE,
  stop_on_error = FALSE
)
}
\arguments{
\item{data}{A data frame of the data file}

\item{meta}{A data frame of the metadata file}

\item{verbose}{logical, if TRUE prints feedback messages to console for
every test, if FALSE run silently}

\item{stop_on_error}{logical, if TRUE will stop with an error if the result
is "FAIL", and will throw genuine warning if result is "WARNING"}
}
\value{
a single row data frame
}
\description{
Validates that every filter column in the data has at least two distinct
values. A filter with only one level provides no analytical value and
should be removed from the metadata.
}
\examples{
precheck_filter_not_singular(example_data, example_meta)
precheck_filter_not_singular(example_data, example_meta, verbose = TRUE)
}
\concept{precheck_filter}
