% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/precheck_geog_level.R
\name{precheck_geog_level}
\alias{precheck_geog_level}
\title{Check that the geographic_level values are all valid.}
\usage{
precheck_geog_level(data, verbose = FALSE, stop_on_error = FALSE)
}
\arguments{
\item{data}{A data frame of the data file}

\item{verbose}{logical, if TRUE prints feedback messages to console for
every test, if FALSE run silently}

\item{stop_on_error}{logical, if TRUE will stop with an error if the result
is "FAIL", and will throw genuine warning if result is "WARNING"}
}
\value{
a single row data frame
}
\description{
Throw an error if geographic_level values used in the data file are not
present in the list of acceptable geographic levels
}
\examples{
precheck_geog_level(example_data)
precheck_geog_level(example_data, verbose = TRUE)
}
\seealso{
Other precheck_geog: 
\code{\link{precheck_geog_level_present}()}
}
\concept{precheck_geog}
