% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/precheck_meta_col_type.R
\name{precheck_meta_col_type}
\alias{precheck_meta_col_type}
\title{Check col_type entries are valid}
\usage{
precheck_meta_col_type(meta, output = "console")
}
\arguments{
\item{meta}{A data frame of the metadata file}

\item{output}{"table", "error-only", or "console". Default is "console"}
}
\value{
a console message with feedback from the check or a single row data
frame if verbose = FALSE
}
\description{
Make sure that all col_type values are either 'Filter' or 'Indicator'
}
\examples{
precheck_meta_col_type(example_meta)
precheck_meta_col_type(example_meta, output = "table")
}
\seealso{
Other precheck_meta: 
\code{\link{precheck_meta_ob_unit}()}
}
\concept{precheck_meta}
