% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/precheck_meta_ob_unit.R
\name{precheck_meta_ob_unit}
\alias{precheck_meta_ob_unit}
\title{Check there are no observational units with rows in the metadata}
\usage{
precheck_meta_ob_unit(meta, verbose = FALSE, stop_on_error = FALSE)
}
\arguments{
\item{meta}{A data frame of the metadata file}

\item{verbose}{logical, if TRUE prints feedback messages to console for
every test, if FALSE run silently}

\item{stop_on_error}{logical, if TRUE will stop with an error if the result
is "FAIL", and will throw genuine warning if result is "WARNING"}
}
\value{
a single row data frame
}
\description{
Make sure that no observational unit columns are referenced in the col_name
and filter_grouping_column columns.
}
\examples{
precheck_meta_ob_unit(example_meta)
precheck_meta_ob_unit(example_meta, verbose = TRUE)
}
\seealso{
Other precheck_meta: 
\code{\link{precheck_meta_col_name}()},
\code{\link{precheck_meta_col_type}()}
}
\concept{precheck_meta}
