% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/precheck_time_id_mix.R
\name{precheck_time_id_mix}
\alias{precheck_time_id_mix}
\title{Check the mix of time identifiers in the data file}
\usage{
precheck_time_id_mix(data, verbose = FALSE, stop_on_error = FALSE)
}
\arguments{
\item{data}{A data frame of the data file}

\item{verbose}{logical, if TRUE prints feedback messages to console for
every test, if FALSE run silently}

\item{stop_on_error}{logical, if TRUE will stop with an error if the result
is "FAIL", and will throw genuine warning if result is "WARNING"}
}
\value{
a single row data frame
}
\description{
Check that only compatible time identifiers are together in the same file.
}
\examples{
precheck_time_id_mix(example_data)
}
\seealso{
Other precheck_time: 
\code{\link{precheck_time_id_valid}()}
}
\concept{precheck_time}
