% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reference_values.R
\docType{data}
\name{req_data_cols}
\alias{req_data_cols}
\title{Required data columns}
\format{
A vector of the required data column names
}
\source{
maintained by explore.statistics@education.gov.uk
}
\usage{
req_data_cols
}
\description{
Required data columns
}
\keyword{datasets}
