% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/req_acceptable_values.R
\docType{data}
\name{req_meta_cols}
\alias{req_meta_cols}
\title{Required metadata columns}
\format{
A vector of the required metadata column names
}
\source{
maintained by explore.statistics@education.gov.uk
}
\usage{
req_meta_cols
}
\description{
Required metadata columns
}
\keyword{datasets}
