% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/screen_dfs.R
\name{screen_dfs}
\alias{screen_dfs}
\title{Run all checks against data and metadata}
\usage{
screen_dfs(data, meta, output = "table")
}
\arguments{
\item{data}{data.frame, for the data table, more efficient if supplied as a
lazy duckplyr data.frame}

\item{meta}{data.frame, for the metadata table}

\item{output}{Control the format of output, either 'table', 'error-only', or
'console'}
}
\value{
data.frame containing the results of the screening
}
\description{
Run all of the checks from the package against the data and metadata
objects.
}
\details{
Provide a pair of data.frames and this will run through the checks in order.
}
\examples{
screen_dfs(example_data, example_meta)
}
