% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/screen_filenames.R
\name{screen_filenames}
\alias{screen_filenames}
\title{Run all checks against filenames}
\usage{
screen_filenames(datafilename, metafilename, output = "console")
}
\arguments{
\item{datafilename}{Character string, name of the data file}

\item{metafilename}{Character string, name of the metadata file}

\item{output}{Control the format of output, either 'table', 'error-only', or
'console'}
}
\value{
data.frame containing the results of the screening
}
\description{
Run all of the checks from the package that relate to the names of the files
themselves.
}
\details{
Provide a pair of filenames, e.g. "myfile.csv" and "myfile.meta.csv" and it
will check they follow the filename conventions.
}
\examples{
screen_filenames("myfile.csv", "myfile.meta.csv")
screen_filenames("myfile.csv", "mymeta.csv", output = "table")
}
