% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_output.R
\name{test_output}
\alias{test_output}
\title{Create test output}
\usage{
test_output(test_name, result, message, guidance_url = NA, console = TRUE)
}
\arguments{
\item{test_name}{name of the test being run}

\item{result}{result of the test being run}

\item{message}{message to be returned}

\item{guidance_url}{optional url for more information}

\item{console}{(optional) if FALSE, will return a single row data frame for
combining multiple test outputs, otherwise prints feedback messages to
console, default is TRUE}
}
\value{
messages to the console giving feedback to the user or a single
row data frame if console = FALSE
}
\description{
Helper function to handle the output from an individual test, gives us an
easier way to make changes in the future if we want to change the output
format.
}
\details{
Made available to users for documentation purposes to illustrate the output
format of the checks.
}
