% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_col_names_spaces.R
\name{check_col_names_spaces}
\alias{check_col_names_spaces}
\title{Check for spaces in variable names
This function checks for spaces in the variable names of a given data frame.}
\usage{
check_col_names_spaces(data, verbose = FALSE, stop_on_error = FALSE)
}
\arguments{
\item{data}{A data frame to be checked for spaces in variable names.}

\item{verbose}{Logical, if TRUE prints feedback messages to console for
every test, if FALSE run silently. Default is FALSE.}

\item{stop_on_error}{Logical, if TRUE will stop with an error if the result
is "FAIL", and will throw genuine warning if result is "WARNING". Default is FALSE.}
}
\value{
a single row data frame indicating the result of the check.
}
\description{
Check for spaces in variable names
This function checks for spaces in the variable names of a given data frame.
}
\examples{
check_col_names_spaces(example_data)
}
\concept{check_col_names}
