% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_empty_rows.R
\name{check_empty_rows}
\alias{check_empty_rows}
\title{Check for empty rows in data}
\usage{
check_empty_rows(data, file_type = NULL)
}
\arguments{
\item{data}{A data frame to be checked for empty rows.}

\item{file_type}{An optional string indicating the type of file being
checked. This will be included in the output check name.}
}
\value{
A list containing:
\describe{
\item{check}{Character string with the stage that the check belongs to.}
\item{result}{Character string, either "PASS", "FAIL" or "ADVISORY".}
\item{message}{Character string with feedback about the check.}
}
}
\description{
This function checks for empty rows in a given data set and returns a list
with the check result and a message.
}
\examples{
data <- data.frame(a = c(1, 2, NA, 4), b = c(NA, 2, 3, 4))
check_empty_rows(data)

}
