% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_filename_spaces.R
\name{check_filename_spaces}
\alias{check_filename_spaces}
\title{Check for spaces in filename}
\usage{
check_filename_spaces(filename, custom_name = NULL, verbose = TRUE)
}
\arguments{
\item{filename}{A character string of the filename to check}

\item{custom_name}{An optional character string to customize the test name
in the output}

\item{verbose}{Logical, if TRUE then messages are printed to the console,
otherwise a data frame is returned so multiple results can be combined into
a single table. Default is TRUE}
}
\value{
the output from \code{test_output()}
}
\description{
This function checks if the provided filename contains any spaces.
}
\examples{
check_filename_spaces("datafile.csv")
check_filename_spaces("data file.csv", verbose = FALSE)
check_filename_spaces("datafile.meta.csv", custom_name = "meta")
}
