% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_filename_spaces.R
\name{check_filename_spaces}
\alias{check_filename_spaces}
\title{Check for spaces in filename}
\usage{
check_filename_spaces(filename, custom_name = NULL, output = "console")
}
\arguments{
\item{filename}{A character string of the filename to check}

\item{custom_name}{An optional character string to customize the test name
in the output}

\item{output}{"table", "error-only", or "console". Default is "console"}
}
\value{
a console message with feedback from the check or a single row data
frame if verbose = FALSE
}
\description{
This function checks if the provided filename contains any spaces.
}
\examples{
check_filename_spaces("datafile.csv")
check_filename_spaces("data file.csv", output = "table")
check_filename_spaces("datafile.meta.csv", custom_name = "meta")
}
\seealso{
Other filename: 
\code{\link{check_filename_special}()},
\code{\link{check_filenames_match}()}
}
\concept{filename}
