% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_filename_spaces.R
\name{check_filename_spaces}
\alias{check_filename_spaces}
\title{Check for spaces in filename}
\usage{
check_filename_spaces(filename, file_type)
}
\arguments{
\item{filename}{A character string representing the data filename.}

\item{file_type}{A character string representing the type of file, usually
either "data" or "metadata"}
}
\value{
A list containing:
\describe{
\item{result}{Character string, either "PASS", "FAIL" or "ADVISORY".}
\item{message}{Character string with feedback about the check.}
\item{check}{Character string with the stage that the check belongs to.}
}
}
\description{
This function checks if the provided filename contains any spaces.
}
\examples{
check_filename_spaces("datafile.csv", "data")
check_filename_spaces("data file.csv", "data")
check_filename_spaces("data_file.meta.csv", "metadata")
}
