% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_filename_special.R
\name{check_filename_special}
\alias{check_filename_special}
\title{Check for special characters in filename}
\usage{
check_filename_special(
  filename,
  custom_name = NULL,
  verbose = FALSE,
  stop_on_error = FALSE
)
}
\arguments{
\item{filename}{A character string of the filename to check}

\item{custom_name}{An optional character string to customize the test name
in the output}

\item{verbose}{logical, if TRUE prints feedback messages to console for
every test, if FALSE run silently}

\item{stop_on_error}{logical, if TRUE will stop with an error if the result
is "FAIL", and will throw genuine warning if result is "WARNING"}
}
\value{
a single row data frame
}
\description{
This function checks if the provided filename contains any special
characters.
}
\examples{
check_filename_special("datafile.csv")
check_filename_special("data+file.meta.csv", verbose = TRUE)
check_filename_special("datafile.meta.csv", custom_name = "meta")
}
\seealso{
Other filename: 
\code{\link{check_filename_spaces}()},
\code{\link{check_filenames_match}()}
}
\concept{filename}
