% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_filter_defaults.R
\name{check_filter_defaults}
\alias{check_filter_defaults}
\title{Check filenames line up between data and metadata files}
\usage{
check_filter_defaults(data, meta, verbose = FALSE, stop_on_error = FALSE)
}
\arguments{
\item{data}{A character string of the data filename to check}

\item{meta}{A character string of the metadata filename to check}

\item{verbose}{logical, if TRUE prints feedback messages to console for
every test, if FALSE run silently}

\item{stop_on_error}{logical, if TRUE will stop with an error if the result
is "FAIL", and will throw genuine warning if result is "WARNING"}
}
\value{
a single row data frame
}
\description{
Making sure the data and metadata file follow the pattern of:
\itemize{
\item datafile.csv (data file)
\item datafile.meta.csv (metadata file)
}
}
\examples{
check_filter_defaults(example_data, example_meta)
}
\seealso{
Other filename: 
\code{\link{check_filename_spaces}()},
\code{\link{check_filename_special}()},
\code{\link{check_filenames_match}()}
}
\concept{filename}
