% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_filter_item_limit.R
\name{check_filter_item_limit}
\alias{check_filter_item_limit}
\title{Check number of filter items are below limit
Ensures that no filter contains more than 25,000 unique options, this
is to protect the EES service against accidental data issues that can
cause performance issues within the admin system.}
\usage{
check_filter_item_limit(
  data,
  meta,
  verbose = FALSE,
  stop_on_error = FALSE,
  filter_item_limit = 25000
)
}
\arguments{
\item{data}{A character string of the data filename to check}

\item{meta}{A character string of the metadata filename to check}

\item{verbose}{logical, if TRUE prints feedback messages to console for
every test, if FALSE run silently}

\item{stop_on_error}{logical, if TRUE will stop with an error if the result
is "FAIL", and will throw genuine warning if result is "WARNING"}

\item{filter_item_limit}{The maximum number of unique items allowed in a
single filter. Default as used by the screener: 25000}
}
\value{
a single row data frame
}
\description{
Check number of filter items are below limit
Ensures that no filter contains more than 25,000 unique options, this
is to protect the EES service against accidental data issues that can
cause performance issues within the admin system.
}
\examples{
check_filter_item_limit(example_data, example_meta)
}
\seealso{
Other check_filter: 
\code{\link{check_filter_blanks}()},
\code{\link{check_filter_defaults}()},
\code{\link{check_filter_group_level}()},
\code{\link{check_filter_ob_total}()},
\code{\link{check_filter_whitespace}()}
}
\concept{check_filter}
