% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_geog_overcompleted_cols.R
\name{check_geog_overcompleted_cols}
\alias{check_geog_overcompleted_cols}
\title{Check for geographic columns completed for unexpected rows}
\usage{
check_geog_overcompleted_cols(
  data,
  meta,
  verbose = FALSE,
  stop_on_error = FALSE
)
}
\arguments{
\item{data}{A data frame of the data file}

\item{meta}{A data frame of the metadata file}

\item{verbose}{logical, if TRUE prints feedback messages to console for
every test, if FALSE run silently}

\item{stop_on_error}{logical, if TRUE will stop with an error if the result
is "FAIL", and will throw genuine warning if result is "WARNING"}
}
\value{
a single row data frame
}
\description{
Checks that geographic columns (code, name, secondary code) are only
populated for rows where the geographic_level is compatible with that
column. For example, \code{region_code} should not contain values for
\code{"National"} rows.
}
\details{
National columns (country_code, country_name) are never checked as they
are expected at all levels. For School and Provider levels, if the name
column is the only filter in the metadata, the name column is skipped as
it is expected to be populated for all rows.
}
\examples{
check_geog_overcompleted_cols(example_data, example_meta)
check_geog_overcompleted_cols(example_data, example_meta, verbose = TRUE)
}
\seealso{
Other check_geog: 
\code{\link{check_geog_country_combos}()},
\code{\link{check_geog_eda_combos}()},
\code{\link{check_geog_ignored_rows}()},
\code{\link{check_geog_la_combos}()},
\code{\link{check_geog_lad_combos}()},
\code{\link{check_geog_lep_combos}()},
\code{\link{check_geog_lsip_combos}()},
\code{\link{check_geog_na}()},
\code{\link{check_geog_na_code}()},
\code{\link{check_geog_other_dupes}()},
\code{\link{check_geog_pcon_combos}()},
\code{\link{check_geog_region_combos}()},
\code{\link{check_geog_region_for_la}()},
\code{\link{check_geog_region_for_lad}()},
\code{\link{check_geog_ward_combos}()}
}
\concept{check_geog}
