% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reference_values.R
\docType{data}
\name{data_dictionary}
\alias{data_dictionary}
\title{Data dictionary}
\format{
A data.frame with columns:
\describe{
\item{col_name}{Standard column names}
\item{col_type}{Column types of defined columns: "Filter" or "Indicator"}
\item{filter_item}{Possible filter items within standard filters}
\item{col_name_parent}{Standardised parent (filter grouping) column names}
\item{filter_item_parent}{Possible parent (filter grouping) filter items}
\item{cbds_code}{Common basic data set code where applicable}
\item{item_error_flag}{Flag as to whether this applies to API data sets or all data sets}
\item{contexts}{Context or publication area of definition}
}
}
\source{
maintained by explore.statistics@education.gov.uk
}
\usage{
data_dictionary
}
\description{
Data dictionary
}
\keyword{datasets}
