% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eesyscreener-package.R
\docType{package}
\name{eesyscreener-package}
\alias{eesyscreener}
\alias{eesyscreener-package}
\title{eesyscreener: Explore education statistics data screening}
\description{
Quality assurance checks used to enforce standards of official statistics files ready for publication on the explore education statistics platform.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://dfe-analytical-services.github.io/eesyscreener/}
  \item \url{https://github.com/dfe-analytical-services/eesyscreener}
  \item Report bugs at \url{https://github.com/dfe-analytical-services/eesyscreener/issues}
}

}
\author{
\strong{Maintainer}: Cam Race \email{cameron.race@education.gov.uk}

Other contributors:
\itemize{
  \item Department for Education, England \email{explore.statistics@education.gov.uk} [copyright holder]
}

}
\keyword{internal}
