% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_datasets.R
\docType{data}
\name{example_comma_data}
\alias{example_comma_data}
\title{Example data that contains strings containing commas beyond the default sniffing length.
The original can be obtained from the EES test suite}
\format{
A data frame with 22,005 rows and 11 columns:
\describe{
\item{time_period}{Time period of row}
\item{time_identifier}{Time identifier of row}
\item{geographic_level}{Geographic level}
\item{country_code}{Country code}
\item{country_name}{Country name}
\item{new_la_code}{New LA code}
\item{la_name}{LA name}
\item{old_la_code}{Old LA code}
\item{sex}{Sex}
\item{education_phase}{Education phase}
\item{enrolment_count}{Enrolment count}
}
}
\source{
maintained by explore.statistics@education.gov.uk
}
\usage{
example_comma_data
}
\description{
Example data that contains strings containing commas beyond the default sniffing length.
The original can be obtained from the EES test suite
}
\keyword{datasets}
