% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_duplicate_rows.R
\name{find_duplicate_rows}
\alias{find_duplicate_rows}
\title{Identify duplicate rows}
\usage{
find_duplicate_rows(data, meta)
}
\description{
Identify duplicate rows in a data set
}
\examples{
find_duplicate_rows(
  dplyr::bind_rows(example_data, example_data),
  example_meta
)
find_duplicate_rows(
  dplyr::bind_rows(example_data, example_data |> dplyr::mutate(enrolment_count = 5)),
  example_meta
)

}
