% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/precheck_col_req_data.R
\name{precheck_col_req_data}
\alias{precheck_col_req_data}
\title{Check all required columns are present in data}
\usage{
precheck_col_req_data(data, verbose = FALSE, stop_on_error = FALSE)
}
\arguments{
\item{data}{A data frame of the data file}

\item{verbose}{logical, if TRUE prints feedback messages to console for
every test, if FALSE run silently}

\item{stop_on_error}{logical, if TRUE will stop with an error if the result
is "FAIL", and will throw genuine warning if result is "WARNING"}
}
\value{
a single row data frame
}
\description{
Using the \code{req_data_cols} object.
}
\examples{
precheck_col_req_data(example_data)
precheck_col_req_data(example_data, verbose = TRUE)

}
\seealso{
Other precheck_col: 
\code{\link{precheck_col_invalid_meta}()},
\code{\link{precheck_col_name_duplicate}()},
\code{\link{precheck_col_req_meta}()},
\code{\link{precheck_col_to_rows}()}
}
\concept{precheck_col}
