% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/precheck_col_req_meta.R
\name{precheck_col_req_meta}
\alias{precheck_col_req_meta}
\title{Check all required columns are present in metadata}
\usage{
precheck_col_req_meta(meta, output = "console")
}
\arguments{
\item{meta}{A data frame of the metadata file}

\item{output}{"table", "error-only", or "console". Default is "console"}
}
\value{
a console message with feedback from the check or a single row data
frame if verbose = FALSE
}
\description{
Using the \code{req_meta_cols} object.
}
\examples{
precheck_col_req_meta(example_meta)
precheck_col_req_meta(example_meta, output = "table")
}
\seealso{
Other precheck_col: 
\code{\link{precheck_col_invalid_meta}()},
\code{\link{precheck_col_req_data}()}
}
\concept{precheck_col}
