% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/precheck_col_to_rows.R
\name{precheck_col_to_rows}
\alias{precheck_col_to_rows}
\title{Quick check of data columns vs metadata rows}
\usage{
precheck_col_to_rows(data, meta, verbose = FALSE, stop_on_error = FALSE)
}
\arguments{
\item{data}{A data frame of the data file}

\item{meta}{A data frame of the metadata file}

\item{verbose}{logical, if TRUE prints feedback messages to console for
every test, if FALSE run silently}

\item{stop_on_error}{logical, if TRUE will stop with an error if the result
is "FAIL", and will throw genuine warning if result is "WARNING"}
}
\value{
a single row data frame
}
\description{
There are 5 mandatory observational unit columns in every data file, so
there should always be at least 5 more columns in the data file than there
are rows in the metadata file.
}
\examples{
precheck_col_to_rows(example_data, example_meta)
precheck_col_to_rows(example_data, example_meta, verbose = TRUE)
}
\seealso{
Other precheck_col: 
\code{\link{precheck_col_invalid_meta}()},
\code{\link{precheck_col_req_data}()},
\code{\link{precheck_col_req_meta}()}
}
\concept{precheck_col}
