% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/precheck_geography_level_present.R
\name{precheck_geography_level_present}
\alias{precheck_geography_level_present}
\title{Check we have the right columns for the geographic level}
\usage{
precheck_geography_level_present(data, verbose = FALSE, stop_on_error = FALSE)
}
\arguments{
\item{data}{A data frame of the data file}

\item{verbose}{logical, if TRUE prints feedback messages to console for
every test, if FALSE run silently}

\item{stop_on_error}{logical, if TRUE will stop with an error if the result
is "FAIL", and will throw genuine warning if result is "WARNING"}
}
\value{
a single row data frame
}
\description{
Check we have the right columns for the geographic level
}
\examples{
precheck_geography_level_present(example_data)
precheck_geography_level_present(example_data, verbose = TRUE)
precheck_geography_level_present(example_comma_data, verbose = TRUE)
}
\seealso{
Other precheck_geog: 
\code{\link{precheck_geog_level}()}
}
\concept{precheck_geog}
