% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/precheck_meta_col_name.R
\name{precheck_meta_col_name}
\alias{precheck_meta_col_name}
\title{Check col_name is completed for all rows}
\usage{
precheck_meta_col_name(meta, output = "console")
}
\arguments{
\item{meta}{A data frame of the metadata file}

\item{output}{"table", "error-only", or "console". Default is "console"}
}
\value{
a console message with feedback from the check or a single row data
frame if verbose = FALSE
}
\description{
Ensure there are no blank cells for col_name in the metafile.
}
\examples{
precheck_meta_col_name(example_meta)
precheck_meta_col_name(example_meta, output = "table")
}
\seealso{
Other precheck_meta: 
\code{\link{precheck_meta_col_type}()},
\code{\link{precheck_meta_ob_unit}()}
}
\concept{precheck_meta}
