% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/precheck_meta_ob_unit.R
\name{precheck_meta_ob_unit}
\alias{precheck_meta_ob_unit}
\title{Check there are no observational units with rows in the metadata}
\usage{
precheck_meta_ob_unit(meta, output = "console")
}
\arguments{
\item{meta}{A data frame of the metadata file}

\item{output}{"table", "error-only", or "console". Default is "console"}
}
\value{
a console message with feedback from the check or a single row data
frame if verbose = FALSE
}
\description{
Make sure that no observational unit columns are referenced in the col_name
and filter_grouping_column columns.
}
\examples{
precheck_meta_ob_unit(example_meta)
precheck_meta_ob_unit(example_meta, output = "table")
}
\seealso{
Other precheck_meta: 
\code{\link{precheck_meta_col_name}()},
\code{\link{precheck_meta_col_type}()}
}
\concept{precheck_meta}
