% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/precheck_time_id_valid.R
\name{precheck_time_id_valid}
\alias{precheck_time_id_valid}
\title{Check all time_identifier values are valid}
\usage{
precheck_time_id_valid(data, meta, output = "console")
}
\arguments{
\item{data}{A data frame of the data file}

\item{meta}{A data frame of the metadata file}

\item{output}{"table", "error-only", or "console". Default is "console"}
}
\value{
a console message with feedback from the check or a single row data
frame if verbose = FALSE
}
\description{
Checks against the \code{acceptable_time_identifiers} object.
}
\examples{
precheck_time_id_valid(example_data, example_meta)
precheck_time_id_valid(example_data, example_meta, output = "table")
}
\concept{precheck_time}
