% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/screen_csv.R
\name{screen_csv}
\alias{screen_csv}
\title{Run all checks against files}
\usage{
screen_csv(
  datapath,
  metapath,
  datafilename = NULL,
  metafilename = NULL,
  output = "table"
)
}
\arguments{
\item{datapath}{Path to the data CSV file}

\item{metapath}{Path to the meta CSV file}

\item{datafilename}{Optional - the name of the data file, if not given it
will be assumed from the path}

\item{metafilename}{Optional - the name of the metadata file, if not given
it will be assumed from the path}

\item{output}{Control the format of output, either 'table', 'error-only', or
'console'}
}
\value{
Console messages or a list containing
\enumerate{
\item A table with the full results of the checks with four columns:
\itemize{
\item result of the check (PASS / FAIL / ADVISORY)
\item message giving feedback about the check
\item group that the check belongs to
\item name of the check
}
\item Overall stage the checks reached
\item Overall message to give back to the user
}
}
\description{
Run all of the checks in the eesyscreener package against your data file and
associated metadata file. It takes in the paths to the CSV files then reads
and parses the files and runs the checks in multiple stages, the function
will return early if any check in a stage fails.
}
\examples{
# Create temp files for the example
data_path <- paste0(tempdir(), "\\\\example.csv")
meta_path <- paste0(tempdir(), "\\\\example.meta.csv")

data.table::fwrite(example_data, data_path)
data.table::fwrite(example_meta, meta_path)

screen_csv(data_path, meta_path)

screen_csv(
  data_path,
  meta_path,
  "data.csv",
  "data.meta.csv",
  output = "console"
)

# Clean up temp files
file.remove(data_path)
file.remove(meta_path)
}
