% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/screen_dfs.R
\name{screen_dfs}
\alias{screen_dfs}
\title{Run all checks against data and metadata}
\usage{
screen_dfs(
  data,
  meta,
  verbose = FALSE,
  stop_on_error = FALSE,
  prudence = "lavish"
)
}
\arguments{
\item{data}{data.frame, for the data table, more efficient if supplied as a
lazy duckplyr data.frame}

\item{meta}{data.frame, for the metadata table}

\item{verbose}{logical, if TRUE prints feedback messages to console for
every test, if FALSE run silently}

\item{stop_on_error}{logical, if TRUE will stop with an error if the result
is "FAIL", and will throw genuine warning if result is "WARNING"}

\item{prudence}{prudence as used by duckplyr, default = "lavish". Can also
be "stingy" and "thrifty".}
}
\value{
data.frame containing the results of the screening
}
\description{
Run all of the checks from the package against the data and metadata
objects.
}
\details{
Provide a pair of data.frames and this will run through the checks in order.
}
\examples{
screen_dfs(example_data, example_meta)
}
