% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/screen_files.R
\name{screen_files}
\alias{screen_files}
\title{Screen files}
\usage{
screen_files(
  datafilename,
  metafilename,
  dataseparator,
  metaseparator,
  data_character,
  meta_character,
  datafile,
  metafile,
  api_only = FALSE
)
}
\arguments{
\item{api_only}{Logical value indicating whether to run only the API checks.}
}
\value{
A table containing the full results of the checks with four columns:
\itemize{
  \item result of the check (PASS / FAIL / ADVISORY)
  \item message giving feedback about the check
  \item stage that the check belongs to
  \item name of the check
}
}
\description{
Run all of the checks in the eesyscreener package against your data file and
associated metadata file. It parses the files and runs the checks in four
stages, the function will return early if any check in a stage fails.
}
\examples{
eesyscreener::screen_files(
  eesyscreener::example_data,
  eesyscreener::example_metadata
)
}
