% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/screen_files.R
\name{screen_files}
\alias{screen_files}
\title{Run all checks against files}
\usage{
screen_files(datafilename, metafilename, datafile, metafile, api_only = FALSE)
}
\arguments{
\item{datafilename}{The name of the data file.}

\item{metafilename}{The name of the metadata file.}

\item{datafile}{A data frame containing the data file contents.}

\item{metafile}{A data frame containing the metadata file contents.}

\item{api_only}{Logical value indicating whether to run only the API checks.}
}
\value{
A list containing
\enumerate{
\item A table with the full results of the checks with four columns:
\itemize{
\item result of the check (PASS / FAIL / ADVISORY)
\item message giving feedback about the check
\item stage that the check belongs to
\item name of the check
}
\item Overall stage the checks reached
\item Overall message to give back to the user
}
}
\description{
Run all of the checks in the eesyscreener package against your data file and
associated metadata file. It parses the files and runs the checks in four
stages, the function will return early if any check in a stage fails.
}
\examples{
screen_files(
  "data.csv",
  "data.meta.csv",
  example_data,
  example_meta
)
}
